+++ QueryBuilder.Tests/Infrastructure/TestCompilers/TestEmptyIdentifiersCompiler.cs
@@ -1,0 +1,7 @@
+﻿namespace SqlKata.Tests.Infrastructure.TestCompilers;
+
+class TestEmptyIdentifiersCompiler : TestCompiler
+{
+    protected override string OpeningIdentifier { get; set; } = "";
+    protected override string ClosingIdentifier { get; set; } = "";
+}
